requirements.txt
absl-py==0.7.1
actionlib==1.11.13
angles==1.9.11
apptools==4.4.0
astor==0.8.0
attrs==19.1.0
autolab-core==0.0.12
-e git+https://github.com/ececyh/robosuite-gqcnn.git@16c240b993df7d9cbeebcb6eed19a6d7db66020b#egg=autolab_perception&subdirectory=robosuite/robosuite/perception
bitarray==1.0.1
bleach==3.1.0
bondpy==1.8.3
Bottleneck==1.2.1
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.18
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
colorlog==4.0.2
configobj==5.0.6
controller-manager==0.13.3
controller-manager-msgs==0.13.3
controller-manager-tests==0.13.3
cv-bridge==1.12.8
cvxopt==1.2.3
cycler==0.10.0
Cython==0.29.10
dask==2.1.0
decorator==4.4.0
defusedxml==0.6.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
dill==0.2.9
dynamic-reconfigure==1.5.50
entrypoints==0.3
envisage==4.7.2
filelock==3.0.12
future==0.17.1
gast==0.2.2
gazebo-plugins==2.5.19
gazebo-ros==2.5.19
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
glfw==1.8.1
google-pasta==0.1.7
GPUtil==1.4.0
-e git+https://github.com/ececyh/robosuite-gqcnn.git@16c240b993df7d9cbeebcb6eed19a6d7db66020b#egg=gqcnn&subdirectory=gqcnn
grpcio==1.21.1
gym==0.13.0
h5py==2.9.0
hidapi==0.7.99.post21
idna==2.8
image-geometry==1.12.8
imageio==2.5.0
interactive-markers==1.11.4
ipykernel==5.1.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
laser-geometry==1.6.4
llvmlite==0.29.0
lockfile==0.12.2
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.0
mayavi==4.7.1
-e git+https://github.com/lianghongzhuo/PointNetGPD.git@ec5cb3d51f9ae1eb5c62fda4f16dff97f53bb44d#egg=meshpy&subdirectory=meshpy
meshrender==0.0.10
message-filters==1.12.14
mistune==0.8.4
mock==3.0.5
moveit-commander==0.9.17
moveit-ros-planning-interface==0.9.17
moveit-ros-visualization==0.9.17
mpmath==1.1.0
msgpack==0.6.1
mujoco-py==2.0.2.2
multiprocess==0.70.7
nbconvert==5.5.0
nbformat==4.4.0
NearPy==1.0.0
networkx==2.3
nose==1.3.7
notebook==5.7.8
numba==0.45.1
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.25.1
pandocfilters==1.4.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluginlib==1.11.3
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.8.0
psutil==5.6.3
ptyprocess==0.6.0
pybullet==1.9.5
pycollada==0.6
pycparser==2.19
pyface==6.1.2
pyglet==1.4.1
Pygments==2.4.2
pyhull==2015.2.1
PyOpenGL==3.1.0
pyparsing==2.4.0
pyquaternion==0.9.5
pyrsistent==0.15.2
pyserial==3.4
python-dateutil==2.8.0
python-pcl==0.3.0a1
python-qt-binding==0.3.4
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.2
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
qtconsole==4.5.1
requests==2.22.0
resource-retriever==1.12.4
-e git+https://github.com/ececyh/robosuite-gqcnn.git@16c240b993df7d9cbeebcb6eed19a6d7db66020b#egg=robosuite&subdirectory=robosuite
rosbag==1.12.14
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.6
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.6
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-joint-trajectory-controller==0.13.5
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
Rtree==0.8.3
ruamel.yaml==0.15.97
rviz==1.12.17
scikit-image==0.14.2
scikit-learn==0.21.2
scikit-video==1.1.11
scipy==1.2.0
Send2Trash==1.5.0
sensor-msgs==1.12.7
setproctitle==1.1.10
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
srdfdom==0.4.2
svg.path==3.0
sympy==1.4
tensorboard==1.12.2
tensorboardX==1.8
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
toolz==0.10.0
topic-tools==1.12.14
torch==1.1.0
tornado==6.0.3
traitlets==4.3.2
traits==5.1.2
traitsui==6.1.2
trimesh==2.38.40
urdfdom-py==0.3.3
urllib3==1.25.3
visualization==0.0.6
visualpriors==0.3
vtk==8.1.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xacro==1.11.3
xxhash==1.3.0
